+++ CHANGELOG.md
@@ -27,11 +27,16 @@
 
 * Add the variables `$pi` and `$e` to the built-in "sass:math" module.
 
+### JavaScript API
+
+* `constructor.value` fields on value objects now match their Node Sass
+  equivalents.
+
 ## 1.24.5
 
 * Highlight contextually-relevant sections of the stylesheet in error messages,
   rather than only highlighting the section where the error was detected.
-
+  
 ## 1.24.4
 
 ### JavaScript API